Pull ump_helpers and string_util into Inc include files
