+++ src/api/Abstraction/JitterReductionAbstraction/Midi2.JitterReductionMidiBidi.cpp
@@ -32,7 +32,7 @@ CMidi2JitterReductionMidiBiDi::Initialize(
 
 
     std::wstring id{ EndpointId };
-    InPlaceToLower(id);
+    Windows::Devices::Midi2::Internal::InPlaceToLower(id);
 
     return S_OK;
 }
