+++ src/api/Inc/ump_helpers.h
@@ -8,8 +8,6 @@
 
 #pragma once
 
-//#include <pch.h>
-
 #define MIDIWORDNIBBLE1(x) ((uint8_t)((x & 0xF0000000) >> 28))
 #define MIDIWORDNIBBLE2(x) ((uint8_t)((x & 0x0F000000) >> 24))
 #define MIDIWORDNIBBLE3(x) ((uint8_t)((x & 0x00F00000) >> 20))
